Rename dataset-generator to example-generator
  ref #6
